Move modules, add error type & fix typos
