+++ crates/full-node/sov-ethereum/src/lib.rs
@@ -76,13 +76,10 @@ where
         handlers::realtime_send_raw_transaction,
     )?;
 
-    rpc.register_async_method(
-        "eth_getLogs",
-        handlers::EthLogsService::<S, Seq>::eth_get_logs,
-    )?;
+    rpc.register_async_method("eth_getLogs", handlers::LogHandlers::<S, Seq>::eth_get_logs)?;
     rpc.register_async_method(
         "eth_getLogsWithCursor",
-        handlers::EthLogsService::<S, Seq>::eth_get_logs_with_cursor,
+        handlers::LogHandlers::<S, Seq>::eth_get_logs_with_cursor,
     )?;
     rpc.register_subscription(
         "eth_subscribe",
